Add second solution based on TOP answer
Created a second solution based on Top solution. The result
in my original one and my solution based on Top's is exactly the
same. I would say that TOP's is better as the position of the
cancel icon (the x) remains in-line with the header text. I rely on
auto margin too much. Use justify-content: space-between more
often when positioning content on the left and right side of a
container.