Add defer attribute to script call in fizzbuzz
I'm adding this so that my html elements load before the script is
called. NOTE THAT (for me anyway) when the .js file was stored in the
browser cache, it is executes immediately despite the defer attribute.
As such, the HTML doesn't load until after the scrip is executed. It
took me a lot of cycles to figure this out :)